corrected output format

+++ equation0.h
@@ -1,11 +1,11 @@
 class equation_0
 {
   private:
-    int equation_term;
-    int equation_result;
+    double equation_term;
+    double equation_result;
 
   public:
-    equation_0(int newEquationTerm, int newEquationResult);
+    equation_0(double newEquationTerm, double newEquationResult);
     ~equation_0();
     void solve();
 };
